Updates UI test code
Adds contexts, re-orders VC code, improves steps (only 1 given, when, then max) etc
